Add files via upload
Used as an automated shell and tube heat exchanger cleaner. X and Y axis orientation stepper motors along. Limit switches used to set the boundaries of the heat exchanger as they come in different sizes. When the Ultrasonic sensor detects by the distance, the hole the motors will stop infront of the hole so that the cleaner can clean then it will detect the next whole. This process will continue until it cleans all the holes and the X-Y axis motors hit the limit switches, indicating it hit the boundaries and thus stopping the operation.